Move ContextMenuItemsGenerationEvent to UI.ContextMenus namespace and remove specific member actions menu injector
